third_party/blink/web_tests/http/tests/serviceworker/service-worker-gc-expected.txt

Test that ServiceWorker and ServiceWorkerRegistration are not garbage collected prematurely

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS registrationObservation.wasCollected is false
PASS swObservation.wasCollected is false
PASS registrationObservation.wasCollected is false
PASS swObservation.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
